[IMPROVEMENT] Move toggle and inquiry to Enterprise
