+++ app/ee/LICENSE
@@ -1,0 +1,43 @@
+The Rocket.Chat Enterprise Edition (EE) license (the "EE License")
+Copyright (c) 2015-2020 Rocket.Chat Technologies Corp.
+
+With regard to the Rocket.Chat Software:
+
+This software and associated documentation files (the "Software") may only be
+used in production, if you (and any entity that you represent) have agreed to,
+and are in compliance with, the Rocket.Chat Subscription Terms of Service,
+available at https://rocket.chat/terms (the "EE Terms"), or other agreement
+governing the use of the Software, as agreed by you and Rocket.Chat, and otherwise
+have a valid Rocket.Chat Enterprise Edition subscription for the correct number
+of user seats. Subject to the foregoing sentence, you are free to modify this
+Software and publish patches to the Software. You agree that Rocket.Chat and/or
+its licensors (as applicable) retain all right, title and interest in and to all
+such modifications and/or patches, and all such modifications and/or patches may
+only be used, copied, modified, displayed, distributed, or otherwise exploited
+with a valid Rocket.Chat Enterprise Edition subscription for the correct number
+of user seats. Notwithstanding the foregoing, you may copy and modify the Software
+for development and testing purposes, without requiring a Subscription. You agree
+that Rocket.Chat and/or its licensors (as applicable) retain all right, title and
+interest in and to all such modifications. You are not granted any other rights
+beyond what is expressly stated herein.  Subject to the foregoing, it is forbidden
+to copy, merge, publish, distribute, sublicense, and/or sell the Software.
+
+This EE License applies only to the part of this Software that is not distributed
+as part of Rocket.Chat Community Edition (CE). Any part of this Software distributed
+as part of Rocket.Chat CE or is served client-side as an image, font, cascading
+stylesheet (CSS), file which produces or is compiled, arranged, augmented, or combined
+into client-side JavaScript, in whole or in part, is copyrighted under the MIT Expat
+license. The full text of this EE License shall be included in all copies or substantial
+portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+
+For all third-party components incorporated into the Rocket.Chat Software,
+those components are licensed under the original license provided by the owner
+of the applicable component.
