+++ app/ee/omnichannel/containers/OmnichannelStatus.js
@@ -7,11 +7,12 @@ import PropTypes from 'prop-types';
 import Touch from '../../../utils/touch';
 import { CustomIcon } from '../../../lib/Icons';
 import I18n from '../../../i18n';
-import styles from '../styles';
+import styles from '../../../views/RoomsListView/styles';
 import { themes, SWITCH_TRACK_COLOR } from '../../../constants/colors';
 import { withTheme } from '../../../theme';
 import UnreadBadge from '../../../presentation/UnreadBadge';
 import RocketChat from '../../../lib/rocketchat';
+import { isOmnichannelStatusAvailable, changeLivechatStatus } from '../lib';
 
 const OmnichannelStatus = memo(({
 	searching, goQueue, theme, queueSize, inquiryEnabled, user
@@ -19,16 +20,16 @@ const OmnichannelStatus = memo(({
 	if (searching > 0 || !(RocketChat.isOmnichannelModuleAvailable() && user?.roles?.includes('livechat-agent'))) {
 		return null;
 	}
-	const [status, setStatus] = useState(user?.statusLivechat === 'available');
+	const [status, setStatus] = useState(isOmnichannelStatusAvailable(user));
 
 	useEffect(() => {
-		setStatus(user.statusLivechat === 'available');
+		setStatus(isOmnichannelStatusAvailable(user));
 	}, [user.statusLivechat]);
 
 	const toggleLivechat = async() => {
 		try {
 			setStatus(v => !v);
-			await RocketChat.changeLivechatStatus();
+			await changeLivechatStatus();
 		} catch {
 			setStatus(v => !v);
 		}
