+++ app/ee/omnichannel/sagas/inquiry.js
@@ -1,19 +1,20 @@
 import { put, takeLatest, select } from 'redux-saga/effects';
 
-import * as types from '../actions/actionsTypes';
-import RocketChat from '../lib/rocketchat';
+import * as types from '../../../actions/actionsTypes';
+import RocketChat from '../../../lib/rocketchat';
 import { inquirySuccess, inquiryFailure, inquirySetEnabled } from '../actions/inquiry';
+import { isOmnichannelStatusAvailable, getInquiriesQueued } from '../lib';
 
 const handleRequest = function* handleRequest() {
 	try {
 		const routingConfig = yield RocketChat.getRoutingConfig();
-		const statusLivechat = yield select(state => state.login.user.statusLivechat);
+		const user = yield select(state => state.login.user);
 		// if routingConfig showQueue is enabled and omnichannel is enabled
-		const showQueue = routingConfig.showQueue && statusLivechat === 'available';
+		const showQueue = routingConfig.showQueue && isOmnichannelStatusAvailable(user);
 
 		if (showQueue) {
 			// get all the current chats on the queue
-			const result = yield RocketChat.getInquiriesQueued();
+			const result = yield getInquiriesQueued();
 			if (result.success) {
 				const { inquiries } = result;
