+++ app/ee/omnichannel/views/QueueListView.js
@@ -4,20 +4,20 @@ import { FlatList } from 'react-native';
 import { connect } from 'react-redux';
 import isEqual from 'react-fast-compare';
 
-import I18n from '../i18n';
-import RoomItem, { ROW_HEIGHT } from '../presentation/RoomItem';
-import { MAX_SIDEBAR_WIDTH } from '../constants/tablet';
-import { isTablet, isIOS } from '../utils/deviceInfo';
-import { getUserSelector } from '../selectors/login';
-import { withTheme } from '../theme';
-import { withDimensions } from '../dimensions';
-import SafeAreaView from '../containers/SafeAreaView';
-import { themes } from '../constants/colors';
-import StatusBar from '../containers/StatusBar';
-import { goRoom } from '../utils/goRoom';
-import { CloseModalButton } from '../containers/HeaderButton';
-import RocketChat from '../lib/rocketchat';
-import { logEvent, events } from '../utils/log';
+import I18n from '../../../i18n';
+import RoomItem, { ROW_HEIGHT } from '../../../presentation/RoomItem';
+import { MAX_SIDEBAR_WIDTH } from '../../../constants/tablet';
+import { isTablet, isIOS } from '../../../utils/deviceInfo';
+import { getUserSelector } from '../../../selectors/login';
+import { withTheme } from '../../../theme';
+import { withDimensions } from '../../../dimensions';
+import SafeAreaView from '../../../containers/SafeAreaView';
+import { themes } from '../../../constants/colors';
+import StatusBar from '../../../containers/StatusBar';
+import { goRoom } from '../../../utils/goRoom';
+import { CloseModalButton } from '../../../containers/HeaderButton';
+import RocketChat from '../../../lib/rocketchat';
+import { logEvent, events } from '../../../utils/log';
 import { getInquiryQueueSelector } from '../selectors/inquiry';
 
 const INITIAL_NUM_TO_RENDER = isTablet ? 20 : 12;
