+++ app/reducers/index.js
@@ -16,7 +16,7 @@ import activeUsers from './activeUsers';
 import usersTyping from './usersTyping';
 import inviteLinks from './inviteLinks';
 import createDiscussion from './createDiscussion';
-import inquiry from './inquiry';
+import inquiry from '../ee/omnichannel/reducers/inquiry';
 import enterpriseModules from './enterpriseModules';
 
 export default combineReducers({
