+++ app/sagas/index.js
@@ -10,7 +10,7 @@ import state from './state';
 import deepLinking from './deepLinking';
 import inviteLinks from './inviteLinks';
 import createDiscussion from './createDiscussion';
-import inquiry from './inquiry';
+import inquiry from '../ee/omnichannel/sagas/inquiry';
 
 const root = function* root() {
 	yield all([
