+++ app/stacks/MasterDetailStack/index.js
@@ -41,7 +41,7 @@ import ScreenLockConfigView from '../../views/ScreenLockConfigView';
 import AdminPanelView from '../../views/AdminPanelView';
 import NewMessageView from '../../views/NewMessageView';
 import CreateChannelView from '../../views/CreateChannelView';
-import QueueListView from '../../views/QueueListView';
+import QueueListView from '../../ee/omnichannel/views/QueueListView';
 import UserPreferencesView from '../../views/UserPreferencesView';
 import UserNotificationPrefView from '../../views/UserNotificationPreferencesView';
