+++ app/views/RoomView/index.js
@@ -55,6 +55,7 @@ import Navigation from '../../lib/Navigation';
 import SafeAreaView from '../../containers/SafeAreaView';
 import { withDimensions } from '../../dimensions';
 import { getHeaderTitlePosition } from '../../containers/Header';
+import { takeInquiry } from '../../ee/omnichannel/lib';
 
 const stateAttrsUpdate = [
 	'joined',
@@ -689,7 +690,7 @@ class RoomView extends React.Component {
 			const { room } = this.state;
 
 			if (this.isOmnichannel) {
-				await RocketChat.takeInquiry(room._id);
+				await takeInquiry(room._id);
 			} else {
 				await RocketChat.joinRoom(this.rid, this.t);
 			}
